Remove unneeded, customized vtkInteractorStyle
